% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterparameter.R
\name{alterParameter}
\alias{alterParameter}
\title{Change the parameters in the previously loaded circuit.}
\usage{
alterParameter(nalter, parameter)
}
\arguments{
\item{nalter}{the total number of times we want to alter the parameter(s).}

\item{parameter}{a list of parameters which will be sent to Spice for
one time's change.}
}
\value{
1. 'Alter command sent to ngspice' if the parameters are altered
successfully; 2. the values of nalter and paramters we sent down.
}
\description{
Change the parameters in the previously loaded circuit.
}
