% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuitload.R
\name{circuitLoad}
\alias{circuitLoad}
\title{Load specified circuit and send it to Spice.}
\usage{
circuitLoad(circarray)
}
\arguments{
\item{circarray}{a list of character strings which are used to define
the circuit.}
}
\value{
the same circuit as the one loaded.
}
\description{
Load specified circuit and send it to Spice.
}
