% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findspice.R
\name{findSpice}
\alias{findSpice}
\title{Find The Path of The Ngspice Shared Library}
\usage{
findSpice(dylibpath, dylibname)
}
\arguments{
\item{dylibpath}{a character string with the general path of the Ngspice shared library. If NULL, the 
function will search across the computer, which is not 
recommended as it may search some directories that requires
root/admin authorization. If a path is specified, the search will be 
conducted under the path recursively with a faster speed.}

\item{dylibname}{a character string with the name of the Ngspice shared library,
the extension of the shared library is added by R automatically.}
}
\value{
The function will stop if no shared library is found, and the first
        available file path containing the shared library will be returned.
}
\description{
Find if the specified Ngspice shared library exists and return the first
 path if there the shared library exists in multiple paths.
}
\examples{
\dontrun{
 # under windows os
  findSpice(dylibpath = 'C:/Users', dylibname = 'ngspice')
 # under *unix os
 findSpice(dylibpath = '/home/user', dylibname = 'libngspice')}
 
}
