% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlength.R
\name{getLength}
\alias{getLength}
\title{Extract the Length of the Output}
\usage{
getLength()
}
\value{
An integer showing the length of the ouput after running a 
        simulation of a circuit in Ngspice.
}
\description{
Extract the length of the output from Ngspice.
}
\examples{
\dontrun{
getLength()
}

}
