% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getplotnames.R
\name{getPlotNames}
\alias{getPlotNames}
\title{Get the all the output names}
\usage{
getPlotNames()
}
\value{
A data fram with two columns: location and Name. The location column
        shows the order of output names obtained from Ngspice and the Name 
        column shows the corresponding output names returned from Ngspice.
}
\description{
\code{getPlotNames} returns the ouput names from Ngspice after running a 
        circuit simulation.
}
