% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializespice.R
\name{initializeSpice}
\alias{initializeSpice}
\title{Initialize NgSpice.}
\usage{
initializeSpice(dllpath, dllname)
}
\arguments{
\item{dllpath}{the path of the dll/so shared library file}

\item{dllname}{the name of the shared library without extensions}
}
\description{
Initialize NgSpice.
}
