% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregulator.R
\docType{data}
\name{linearregulator}
\alias{linearregulator}
\title{Netlist with circuit set up of a linear regulator}
\format{An object of class \code{character} of length 14.}
\usage{
linearregulator
}
\description{
See the vignette of RSpice or Linear Regulator Modified from P. Basso: Switched-Mode Power
 Supplies, 2008, Chapter 1, figure 1-2 for a visualization of the 
 circuit.
}
\keyword{datasets}
