% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opAmp.R
\docType{data}
\name{opAmp}
\alias{opAmp}
\title{Netlist with circuit set up of a negative feedback loop with 741 op-amp}
\format{An object of class \code{character} of length 182.}
\usage{
opAmp
}
\description{
See the vignette of RSpice or Microelectronic circuits by Sedra for more details.
}
\keyword{datasets}
