% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiceCommand.R
\name{spiceCommand}
\alias{spiceCommand}
\title{Change the device or model parameters of the loaded circuit.}
\usage{
spiceCommand(cmd)
}
\arguments{
\item{cmd}{a vector of commands which to be sent to Ngspice for
one time change.}
}
\description{
#Change the device or model parameters of the loaded circuit. See more
details on Chapter 17 and 19 on how to alter the parameters in 
\url{http://ngspice.sourceforge.net/docs/ngspice-manual.pdf}
}
\examples{
\dontrun{spiceCommand("listing")}                 
}
