% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyexample.R
\docType{data}
\name{toyexample}
\alias{toyexample}
\title{A toy example for the use of Ngspice.}
\format{An object of class \code{character} of length 6.}
\usage{
toyexample
}
\description{
A circuit netlist with two resistors and one voltage source connected in 
series. Two resistors have resistance of 5k and 1k Ohm, respectively, with
the voltage scan from -10 to 0 with an increment of 1.
}
\keyword{datasets}
