% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unloadspice.R
\name{unloadSpice}
\alias{unloadSpice}
\title{Unload NgSpice}
\usage{
unloadSpice()
}
\value{
A message "Ngspice shared library unloaded." or "Ngspice shared 
library already unloaded." dependening on if the \code{unloadSpice()} 
function is run for the first time or not.
}
\description{
Unload the ngspice shared library.
}
