% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterparameter.R
\name{alterParameter}
\alias{alterParameter}
\title{Change the device or model parameters of the loaded circuit.}
\usage{
alterParameter(altercmd)
}
\arguments{
\item{altercmd}{a vector of parameters which to be sent to Ngspice for
one time change.}
}
\description{
Change the device or model parameters of the loaded circuit. See more
details on Chapter 17 and 19 on how to alter the parameters in 
\url{http://ngspice.sourceforge.net/docs/ngspice-manual.pdf}
}
\examples{
\dontrun{alterParameter("alter r1=500")}
}
