% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuitload.R
\name{circuitLoad}
\alias{circuitLoad}
\title{Load specified circuit and send it to Ngspice.}
\usage{
circuitLoad(circarray, listing = TRUE)
}
\arguments{
\item{circarray}{a list of character strings which are used to define}

\item{listing}{logical; if TRUE, print a listing of the current circuit after
               loading.
the circuit.}
}
\value{
the same circuit as the one loaded.
}
\description{
Load specified circuit and send it to Ngspice.
}
\examples{
\dontrun{circuitload(c(".title test", "R1 1 2 5k", "R2 2 0 1k",
           "VDD 0 1 DC 10", ".op", ".end"))}
}
