% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuitload.R
\name{circuitLoad}
\alias{circuitLoad}
\title{Load Specified Circuit to Ngspice}
\usage{
circuitLoad(circarray, dylibpath = NULL, dylibname = NULL,
  listing = TRUE, verbose = TRUE)
}
\arguments{
\item{circarray}{a list of character strings which are used to define}

\item{dylibpath}{the path of the ngspice shared library file. If 
\code{dylibpath = NULL}, the program will 
search the default path for the shared library.}

\item{dylibname}{the name of the ngspice shared library without extensions.
If \code{dylibpath = NULL}, the program will search the 
shared library with its default name.}

\item{listing}{logical; if TRUE, print a listing of the current circuit after
loading the circuit.}

\item{verbose}{logical; if TRUE, print the stdout, stderr etc information 
exported from ngspice.}
}
\value{
The outputs from printf, fprintf and fputs of ngspice simulator if 
         \code{verbose = TRUE}. The listing of the current circuit will be
          returned if \code{listing = TRUE}.
}
\description{
Initialize ngspice by linking to its shared library, load the specified
 circuit and send the circuit netlist to ngspice. If the ngspice shared
 library has been linked, only the function will only load the 
 specified circuit.
}
\examples{
\dontrun{
# Load a simple circtui with name test, the resistor R1 is 5k Ohm and 
# connected to node 1 and 2, R2 is 1k Ohm and connected to node 2 and
# 0, and a voltage source VDD with 10V between node 0 and 1. A point 
# analysis is performed.   
circuitload(c('.title test', 'R1 1 2 5k', 'R2 2 0 1k',
           'VDD 0 1 DC 10', '.op', '.end'))}
}
