% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportresults.R
\name{exportResults}
\alias{exportResults}
\title{Export the Output Values from Ngspice}
\usage{
exportResults(location)
}
\arguments{
\item{location}{a vector of the location(s) of the outputs we want to export,
i.e. the location-th output from the \code{getPlotNames()}.}
}
\value{
A matrix containing the output values with the ouputs for each 
        location stored in one row. The number of rows of the
        matrix equals \code{length(location)} and the number of columns
        equals the length of each output value.
}
\description{
Export the specified output values from Ngspice.
}
\examples{
\dontrun{
# obtain the location and Name information by running getPlotNames()
getPlotNames()
# Export the output values for location 1, 2 and 4. 
exportResults(c(1, 2, 4))
}
}
