% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportresults.R
\name{exportResults}
\alias{exportResults}
\title{Export the output from Spice}
\usage{
exportResults(location)
}
\arguments{
\item{location}{a vector of the location(s) of the output we want to export,
i.e. the location-th vector name from the getlength function.}
}
\value{
the list of the output we specified.
}
\description{
Export the output from Spice
}
