% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findspice.R
\name{findSpice}
\alias{findSpice}
\title{Function to find if shared Ngspice library exists and return the path if Yes.}
\usage{
findSpice(dylibpath, dylibname)
}
\arguments{
\item{dylibpath}{the path of the Ngspice shared library.}

\item{dylibname}{the name of the Ngspice shared library.}
}
\value{
The function will stop if no shared library were found, and the first
        available file path containing the shared library will be returned.
}
\description{
Function to find if shared Ngspice library exists and return the path if Yes.
}
