% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findspice.R
\name{findSpice}
\alias{findSpice}
\title{Function to find if shared Ngspice library exists and return the path if Yes.}
\usage{
findSpice(dylibpath, dylibname)
}
\arguments{
\item{dylibpath}{a character string with the general path of the Ngspice shared library. If NULL, the 
function will search across the computer, which is not 
recommended as it may search some directories that requires
root/admin authorization. If a path is specified, the search will be 
conducted under the path recursively.}

\item{dylibname}{a character string with the name of the Ngspice shared library.}
}
\value{
The function will stop if no shared library were found, and the first
        available file path containing the shared library will be returned.
}
\description{
Function to find if shared Ngspice library exists and return the path if Yes.
}
\examples{
\dontrun{findSpice("/path/to/be/searched", "ngspice.dll")}
\dontrun{findSpice("/path/to/be/searched", "libngspice.so")}
}
