% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlength.R
\name{getLength}
\alias{getLength}
\title{Extract the length of the output.}
\usage{
getLength()
}
\value{
An integer showing the length of the ouput after running a 
        simulation on a circuit in Ngspice.
}
\description{
Extract the length of the output.
}
