% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlength.R
\name{getLength}
\alias{getLength}
\title{Extract the length of the output.}
\usage{
getLength()
}
\value{
an integer showing the length of the ouput and a list of output names.
}
\description{
Extract the length of the output.
}
