% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getplotnames.R
\name{getPlotNames}
\alias{getPlotNames}
\title{Read All Output Names from Ngspice}
\usage{
getPlotNames()
}
\value{
A data fram with two columns: location and Name. The location column
        shows the order of output names obtained from Ngspice and the Name 
        column shows the corresponding output names returned from Ngspice.
}
\description{
Read the available output names with their corresponding order from 
Ngspice after the simulation of a circuit.
}
\details{
\code{getPlotNames} returns the output names from Ngspice after running a 
        circuit simulation.
}
\examples{
\dontrun{
getPlotNames()
}
}
