% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getplotnames.R
\name{getPlotNames}
\alias{getPlotNames}
\title{Read All Output Names from Ngspice}
\usage{
getPlotNames()
}
\value{
A data frame with two columns: \code{location} and \code{Name}.
        The \code{location} column shows the order of output names 
        retuned from Ngspice and the \code{Name} 
        column shows the corresponding output names returned from Ngspice.
}
\description{
Read the available output names with their corresponding order from 
Ngspice after the simulation of a circuit.
}
\details{
\code{getPlotNames} returns the output names from Ngspice after running a 
        circuit simulation.
}
\examples{
\dontrun{
getPlotNames() }
}
