% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializespice.R
\name{initializeSpice}
\alias{initializeSpice}
\title{Initialize Ngspice}
\usage{
initializeSpice(dylibpath, dylibname, verbose)
}
\arguments{
\item{dylibpath}{the path of the ngspice shared library file.}

\item{dylibname}{the name of the ngspice shared library without extensions
such as .so or .dll.}
}
\value{
NULL.
}
\description{
Load the ngspice shared library and initialize Ngspice with the exported
 functions. See Chapter 19 of the ngspice manual for more details.
}
\examples{
\dontrun{
# Initialize Ngspice.
initializeSpice(dylibpath = "/usr/local/bin", dylibname = 'libngspice')}
}
