% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializespice.R
\name{initializeSpice}
\alias{initializeSpice}
\title{Initialize Ngspice.}
\usage{
initializeSpice(dylibpath, dylibname)
}
\arguments{
\item{dylibpath}{the path of the .dll/.so Ngspice shared library file}

\item{dylibname}{the name of the shared library without extensions}
}
\description{
Initialize Ngspice.
}
\examples{
\dontrun{initializeSpice(findSpice(), 'libngspice.so')}
}
