% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializespice.R
\name{initializeSpice}
\alias{initializeSpice}
\title{Initialize NgSpice.}
\usage{
initializeSpice(dllpath, dllname)
}
\value{
A character string saying "ngspice loaded".
}
\description{
Initialize NgSpice.
}
