% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runspice.R
\name{runSpice}
\alias{runSpice}
\title{Run Ngspice.}
\usage{
runSpice(bgrun = TRUE)
}
\arguments{
\item{bgrun}{logical; indicator of if the simulation is run in main thread 
or background thread.}
}
\value{
Simulation status obtained from Ngspice is printed.
}
\description{
Run Ngspice.
}
\examples{
\dontrun{
# Run the simulation in main thread
runSpice()

# Run the simulation in background thread 
runSpice(bgrun = TRUE) 
} 
}
