% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runspice.R
\name{runSpice}
\alias{runSpice}
\title{Command to run Ngspice.}
\usage{
runSpice(bgrun = TRUE)
}
\arguments{
\item{bgrun}{logical; indicator of if the simulation is run in main thread 
or background thread.}
}
\description{
Command to run Ngspice.
}
