% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiceCommand.R
\name{spiceCommand}
\alias{spiceCommand}
\title{Send Command to Ngspice}
\usage{
spiceCommand(cmd)
}
\arguments{
\item{cmd}{a vector of commands which to be sent to Ngspice for
one time change.}
}
\description{
Send a valid command from caller
 to ngspice.dll. See more control or interactive commands details in
\url{http://ngspice.sourceforge.net/docs/ngspice-manual.pdf}
}
\examples{
\dontrun{

# to print a listing of the current circuit

spiceCommand("listing")}                 
}
